App is near complete with one problem.
After a user has signed in before, and then signs out--when they sign in again the app signs them into the app that they signed out. The user is unable to sign into a different account.
I do not know how to fix this.

Also need to add title of app to home screen.